Refactor: Split AuthDialog into smaller components
This commit refactors the AuthDialog component into smaller, more manageable components to improve readability and maintainability.
