+++ src/contexts/auth/AuthContext.tsx
@@ -2,4 +2,4 @@
 // This file is a redirection to the new location
 // This helps maintain backward compatibility for any imports still using the old path
 
-export { UserProvider as AuthProvider, useUser as useAuth } from '../user/UserContext';
+export { useUser as useAuth, UserProvider as AuthProvider } from '../user/UserContext';
